+++ diesel/src/doctest_setup.rs
@@ -395,6 +395,7 @@ mod schema {
 
     table! {
         users {
+            #[auto_increment]
             id -> Integer,
             name -> VarChar,
         }